refactor common test parts
